+++ kafka_consumer/datadog_checks/kafka_consumer/kafka_consumer.py
@@ -2,6 +2,8 @@
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 
+import time
+
 from datadog_checks.base import AgentCheck, is_affirmative
 from datadog_checks.kafka_consumer.client.kafka_client_factory import make_client
 
@@ -42,15 +44,15 @@ class KafkaCheck(AgentCheck):
 
         # Fetch Kafka consumer offsets
         try:
-            self.client.get_consumer_offsets()
+            self._consumer_offsets = self.client.get_consumer_offsets()
         except Exception:
             self.log.exception("There was a problem collecting consumer offsets from Kafka.")
             # don't raise because we might get valid broker offsets
 
         # Fetch the broker highwater offsets
         try:
-            if self.client.should_get_highwater_offsets():
-                self.client.get_highwater_offsets()
+            if self.should_get_highwater_offsets():
+                self._highwater_offsets = self.client.get_highwater_offsets()
             else:
                 self.warning("Context limit reached. Skipping highwater offset collection.")
         except Exception:
@@ -69,15 +71,111 @@ class KafkaCheck(AgentCheck):
             )
 
         # Report the metrics
-        self.client.report_highwater_offsets(self._context_limit)
-        self.client.report_consumer_offsets_and_lag(self._context_limit - len(self.client._highwater_offsets))
+        self.report_highwater_offsets(self._context_limit)
+        self.report_consumer_offsets_and_lag(self._context_limit - len(self._highwater_offsets))
 
         self.collect_broker_metadata()
 
+    def report_highwater_offsets(self, contexts_limit):
+        """Report the broker highwater offsets."""
+        reported_contexts = 0
+        self.log.debug("Reporting broker offset metric")
+        for (topic, partition), highwater_offset in self._highwater_offsets.items():
+            broker_tags = ['topic:%s' % topic, 'partition:%s' % partition]
+            broker_tags.extend(self._custom_tags)
+            self.gauge('broker_offset', highwater_offset, tags=broker_tags)
+            reported_contexts += 1
+            if reported_contexts == contexts_limit:
+                return
+
+    def report_consumer_offsets_and_lag(self, contexts_limit):
+        """Report the consumer offsets and consumer lag."""
+        reported_contexts = 0
+        self.log.debug("Reporting consumer offsets and lag metrics")
+        for (consumer_group, topic, partition), consumer_offset in self._consumer_offsets.items():
+            if reported_contexts >= contexts_limit:
+                self.log.debug(
+                    "Reported contexts number %s greater than or equal to contexts limit of %s, returning",
+                    str(reported_contexts),
+                    str(contexts_limit),
+                )
+                return
+            consumer_group_tags = ['topic:%s' % topic, 'partition:%s' % partition, 'consumer_group:%s' % consumer_group]
+            consumer_group_tags.extend(self._custom_tags)
+
+            # TODO: get_partitions_for_topic()
+            partitions = self.client.kafka_client._client.cluster.partitions_for_topic(topic)
+            self.log.debug("Received partitions %s for topic %s", partitions, topic)
+            if partitions is not None and partition in partitions:
+                # report consumer offset if the partition is valid because even if leaderless the consumer offset will
+                # be valid once the leader failover completes
+                self.gauge('consumer_offset', consumer_offset, tags=consumer_group_tags)
+                reported_contexts += 1
+
+                if (topic, partition) not in self._highwater_offsets:
+                    self.log.warning(
+                        "Consumer group: %s has offsets for topic: %s partition: %s, but no stored highwater offset "
+                        "(likely the partition is in the middle of leader failover) so cannot calculate consumer lag.",
+                        consumer_group,
+                        topic,
+                        partition,
+                    )
+                    continue
+                producer_offset = self._highwater_offsets[(topic, partition)]
+                consumer_lag = producer_offset - consumer_offset
+                if reported_contexts < contexts_limit:
+                    self.gauge('consumer_lag', consumer_lag, tags=consumer_group_tags)
+                    reported_contexts += 1
+
+                if consumer_lag < 0:
+                    # this will effectively result in data loss, so emit an event for max visibility
+                    title = "Negative consumer lag for group: {}.".format(consumer_group)
+                    message = (
+                        "Consumer group: {}, topic: {}, partition: {} has negative consumer lag. This should never "
+                        "happen and will result in the consumer skipping new messages until the lag turns "
+                        "positive.".format(consumer_group, topic, partition)
+                    )
+                    key = "{}:{}:{}".format(consumer_group, topic, partition)
+                    self.send_event(title, message, consumer_group_tags, 'consumer_lag', key, severity="error")
+                    self.log.debug(message)
+            else:
+                if partitions is None:
+                    msg = (
+                        "Consumer group: %s has offsets for topic: %s, partition: %s, but that topic has no partitions "
+                        "in the cluster, so skipping reporting these offsets."
+                    )
+                else:
+                    msg = (
+                        "Consumer group: %s has offsets for topic: %s, partition: %s, but that topic partition isn't "
+                        "included in the cluster partitions, so skipping reporting these offsets."
+                    )
+                self.log.warning(msg, consumer_group, topic, partition)
+
+                # TODO: Include request_update() temporarily since kafka-python may require it
+                self.client.kafka_client._client.cluster.request_update()  # force metadata update on next poll()
+
     @AgentCheck.metadata_entrypoint
     def collect_broker_metadata(self):
-        self.client.collect_broker_metadata()
+        version = self.client.get_version()
+        version_data = [str(part) for part in version]
+        version_parts = {name: part for name, part in zip(('major', 'minor', 'patch'), version_data)}
 
-    # TODO: Remove me once the tests are refactored
+        self.set_metadata(
+            'version', '.'.join(version_data), scheme='parts', final_scheme='semver', part_map=version_parts
+        )
+
+    def should_get_highwater_offsets(self):
+        return len(self._consumer_offsets) < self._context_limit
+
     def send_event(self, title, text, tags, event_type, aggregation_key, severity='info'):
-        self.client._send_event(title, text, tags, event_type, aggregation_key, severity='info')
+        """Emit an event to the Datadog Event Stream."""
+        event_dict = {
+            'timestamp': int(time()),
+            'msg_title': title,
+            'event_type': event_type,
+            'alert_type': severity,
+            'msg_text': text,
+            'tags': tags,
+            'aggregation_key': aggregation_key,
+        }
+        self.event(event_dict)
